refactor: move message attachment handling under attach.parts
- move incoming MIME part handling from attach.lua to attach/parts.lua
- update ftplugins to require notmuch.attach.parts
- update attachment part specs for new module path
- rename attach_spec to attach_parts_spec
- keep behavior unchanged while preparing attach namespace refactor
